+++ packages/web-extension/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rrweb/web-extension
 
+## 2.0.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @saola.ai/rrweb-player@2.0.12
+  - @saola.ai/rrweb@2.0.12
+
 ## 2.0.11
 
 ### Patch Changes